Move `message` to dedicated module
