+++ src/device/accel.rs
@@ -1,4 +1,4 @@
-use crate::game::{driver::MAILBOX_CAPACITY, Message};
+use crate::game::{driver::MAILBOX_CAPACITY, message::Message};
 use core::{fmt::Debug, marker::PhantomData};
 use lsm303agr::{
     interface::{I2cInterface, ReadData, WriteData},
